Revert "Revert "Merge branch 'feature/JNG-2630_Col_value_calculation'""
This reverts commit c365c5c78cd89aded0387513a66ef1567ea45fff.

+++ lib/src/judo_switch.dart
@@ -17,7 +17,7 @@ class JudoSwitch extends StatefulWidget {
     this.alignment = Alignment.centerLeft,
   }) : super(key: key);
 
-  final int col;
+  final double col;
   final double row;
   final bool mandatory;
   final String label;
